Wire cache work to transfer request
We deprecate the buf map design. Hence the cache feature is no longer supported in the buf_map type of API. The cache framework is moved to transfer request.
